Rename #update(model, attributes) into #update_attributes(model, attributes)
